requirements.txt
# submission
requests
huggingface_hub

# core dependencies
torch           # install with the appropriate CUDA index URL
torchvision
torchaudio
lightning
hydra-core
hydra-submitit-launcher  # For jean-Zay cluster job submission
omegaconf
wandb
gcsfs
s3fs

# data / scientific
numpy
pandas
xarray
netcdf4
h5netcdf
h5py
zarr<3
einops
scipy
tqdm
rich
tensorstore # For fast zarr reading during training

# Jupyter and related tools
jupyterlab
ipywidgets
ipykernel

# visualization
matplotlib
cartopy
cmocean

# distributed
dask
distributed

# (optional) fast attention implementation
#flash-attn           # only if you want to enable USE_FLASH_ATTN

# testing (also useful for development)
pytest
pytest_cov

# for building documentation
sphinx
sphinx_rtd_theme
sphinx-autodoc-typehints